backend/requirements.txt
aniso8601==9.0.1
asgiref==3.8.1
Django==5.1
django-environ==0.11.2
django-filter==24.3
django-graphql-jwt==0.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
graphene==3.3
graphene-django==3.2.2
graphql-core==3.2.3
graphql-relay==3.2.0
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
promise==2.3
PyJWT==2.9.0
pytest==8.3.2
pytest-django==4.8.0
python-dotenv==1.0.1
six==1.16.0
sqlparse==0.5.1
text-unidecode==1.3
